Tested and done with ticketing system

+++ database/migrations/1616673296705_tickets.ts
@@ -3,11 +3,11 @@ import BaseSchema from '@ioc:Adonis/Lucid/Schema'
 export default class Tickets extends BaseSchema {
   protected tableName = 'tickets'
 
-  public async up () {
+  public async up() {
     this.schema.createTable(this.tableName, (table) => {
       table.increments('id').primary()
       table.string('code')
-      table.float('amount')
+      table.double('amount')
       table.integer('user_id').unsigned().references('id').inTable('users').onDelete('CASCADE')
       table.integer('event_id').unsigned().references('id').inTable('tickets').onDelete('CASCADE')
       table.boolean('is_used').defaultTo(false)
@@ -16,7 +16,7 @@ export default class Tickets extends BaseSchema {
     })
   }
 
-  public async down () {
+  public async down() {
     this.schema.dropTable(this.tableName)
   }
 }
